dashboard-backend/requirements.txt
appdirs==1.4.2
bleach==2.0.0
defusedxml==0.5.0
Django==1.11
django-crontab==0.7.1
django-extensions==1.7.7
django-fernet-fields
django-filter
djangorestframework
djangorestframework-jwt
elasticsearch==5.3.0
filelock==2.0.7
html5lib==0.999999999
jira==1.0.10
jsonfield==2.0.2
markdown
mysqlclient==1.3.10
oauthlib==2.0.2
packaging==16.8
pbr==2.0.0
pyparsing==2.1.10
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
urllib3==1.20
webencodings==0.5